+++ CHANGELOG.md
@@ -1,4 +1,8 @@
 ## [Unreleased]
+### Added
+* Feature::layer and ::num_tags methods
+### Changed
+* Feature::set_id can no longer fail
 
 ## [0.4.0] - 2019-02-07
 ### Added